surface preferences menu in mobile web
