+++ src/components/PageMenu/PageMenuItem.tsx
@@ -5,7 +5,6 @@ import { PageMenuRadioGroupItem } from './PageMenuRadioGroup';
 
 export enum PageMenuItemType {
   Navigation = 'navigation',
-  Toggle = 'toggle',
   RadioGroup = 'radioGroup',
 }
 
@@ -16,9 +15,6 @@ export const PageMenuItem = <MenuItemValue extends string>({
   switch (type) {
     case PageMenuItemType.Navigation:
       return <PageMenuNavigationItem {...props} />;
-    case PageMenuItemType.Toggle:
-      // TODO: implement toggle item component when needed for notifications settings
-      return null;
     case PageMenuItemType.RadioGroup:
       return <PageMenuRadioGroupItem {...props} />;
     default:
