+++ src/constants/routes.ts
@@ -35,7 +35,7 @@ export enum HistoryRoute {
 
 export enum MobileSettingsRoute {
   Language = 'language',
-  Notifications = 'notifications',
+  Preferences = 'preferences',
   Network = 'network',
 }
