+++ src/pages/settings/SettingsHeader.tsx
@@ -24,12 +24,12 @@ export const SettingsHeader = ({
       label: stringGetter({ key: STRING_KEYS.LANGUAGE }),
     },
     {
-      value: `${AppRoute.Settings}/${MobileSettingsRoute.Notifications}`,
-      label: stringGetter({ key: STRING_KEYS.NOTIFICATIONS }),
-    },
-    {
       value: `${AppRoute.Settings}/${MobileSettingsRoute.Network}`,
       label: stringGetter({ key: STRING_KEYS.NETWORK }),
+    },
+    {
+      value: `${AppRoute.Settings}/${MobileSettingsRoute.Preferences}`,
+      label: stringGetter({ key: STRING_KEYS.PREFERENCES }),
     },
   ];
